Consolidate into one file.
