+++ .gitignore
@@ -1,4 +1,4 @@
 /inspect.lua
-/sm64_rando_tracker/*
-!/sm64_rando_tracker/Json.lua
-!/sm64_rando_tracker/logo.png
+/ironmario_tracker/*
+!/ironmario_tracker/Json.lua
+!/ironmario_tracker/logo.png
